+++ MultisigWallet/MultisigWalletApplication/WalletApplicationService.swift
@@ -144,7 +144,7 @@ public class WalletApplicationService: Assertable {
         return findSelectedWallet()?.address?.value
     }
 
-    public var minimumDeploymentAmount: Int? {
+    public var minimumDeploymentAmount: BigInt? {
         return findAccount("ETH")?.minimumDeploymentTransactionAmount
     }
 
@@ -249,8 +249,7 @@ public class WalletApplicationService: Assertable {
                 .contains(selectedWalletState) else {
                 return RepeatingShouldStop.yes
             }
-            // TODO: BigInt support
-            update(account: "ETH", newBalance: Int(newBalance)) // mutates selectedWalletState
+            update(account: "ETH", newBalance: newBalance)
             if selectedWalletState == .accountFunded {
                 try createWalletInBlockchain()
                 return RepeatingShouldStop.yes
@@ -265,7 +264,6 @@ public class WalletApplicationService: Assertable {
 
     private func createWalletInBlockchain() throws {
         let address = findSelectedWallet()!.address!
-        // TODO: if the call fails, show error in UI and possibility to retry with a button
         try ethereumService.startSafeCreation(address: address)
         guard selectedWalletState == .accountFunded else { return }
         markDeploymentAcceptedByBlockchain()
@@ -325,8 +323,7 @@ public class WalletApplicationService: Assertable {
     private func fetchBalance() throws {
         let address = findSelectedWallet()!.address!.value
         let newBalance = try ethereumService.balance(address: address)
-        // TODO: BigInt support
-        update(account: "ETH", newBalance: Int(newBalance))
+        update(account: "ETH", newBalance: newBalance)
     }
 
     private func removePaperWallet() {
@@ -438,7 +435,6 @@ public class WalletApplicationService: Assertable {
             throw Error.validationFailed
         } catch let JSONHTTPClient.Error.networkRequestFailed(_, response, data) {
             if let data = data, let dataStr = String(data: data, encoding: .utf8),
-                // FIXME: fragile error detection, better to have JSON code/struct
                 dataStr.range(of: "Exceeded expiration date") != nil {
                 throw Error.exceededExpirationDate
             } else if let response = response as? HTTPURLResponse {
@@ -472,15 +468,14 @@ public class WalletApplicationService: Assertable {
 
     // MARK: - Accounts
 
-    public func accountBalance(token: String) -> Int? {
+    public func accountBalance(token: String) -> BigInt? {
        return findAccount(token)?.balance
     }
 
     private func updateMinimumFunding(account token: String, amount: BigInt) {
         assertCanChangeAccount()
         mutateAccount(token: token) { account in
-            // TODO: bigint
-            account.updateMinimumTransactionAmount(Int(amount))
+            account.updateMinimumTransactionAmount(amount)
         }
     }
 
@@ -488,7 +483,7 @@ public class WalletApplicationService: Assertable {
         try! assertTrue(selectedWalletState.isValidForAccountUpdate, Error.invalidWalletState)
     }
 
-    public func update(account token: String, newBalance: Int) {
+    public func update(account token: String, newBalance: BigInt) {
         assertCanChangeAccount()
         mutateAccount(token: token) { account in
             account.update(newAmount: newBalance)
